+++ test/plugins/test_shares.py
@@ -51,7 +51,8 @@ class TestShares:
         class TestSharesPlugin(SharesPlugin):
 
             def _get_table_rows(self, t_data: str) -> Generator[Dict, None, None]:
-                return utils._safe_get_fixture_entries(TestShares.FIXTURES, t_data, limit=2)
+                limit = 3 if t_data == "APP.V_INBOUND_SHARE_TABLES" else 2
+                return utils._safe_get_fixture_entries(TestShares.FIXTURES, t_data, limit=limit)
 
         def __local_get_plugin_class(source: str):
             return TestSharesPlugin
@@ -77,7 +78,7 @@ class TestShares:
                 affecting_types_for_entries=["logs", "events"],  # there is not test data for events
                 base_count={
                     "outbound_shares": {"entries": 2, "log_lines": 2, "metrics": 0, "events": 2},
-                    "inbound_shares": {"entries": 2, "log_lines": 2, "metrics": 0, "events": 0},
+                    "inbound_shares": {"entries": 3, "log_lines": 3, "metrics": 0, "events": 0},
                     "shares": {"entries": 2, "log_lines": 0, "metrics": 0, "events": 2},
                 },
             )